requirements.txt
# RLC-Agent Dependencies
# Install with: uv pip install -r requirements.txt

# Core dependencies
aiohttp>=3.9.0
httpx>=0.27.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL
pymysql>=1.1.0          # MySQL

# API clients
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Scheduling
apscheduler>=3.10.0

# Office document generation
python-docx>=1.0.0

# Cloud integrations (optional)
dropbox>=11.0.0

# Configuration
pyyaml>=6.0      # YAML config files (sensitivity, etc.)

# Data processing
openpyxl>=3.1.0  # Excel files
pypdf>=4.0.0     # PDF parsing for RAG

# Web search (for agent tools)
tavily-python>=0.3.0  # Primary search backend
ddgs>=6.0.0           # Fallback (renamed from duckduckgo-search)

# Google API (for email and calendar)
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0

# Notion API (for long-term memory)
notion-client>=2.0.0

# Local embeddings (for RAG)
sentence-transformers>=2.2.0  # Optional - for enhanced RAG

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
